OpenPaaS-Suite/esn#10 Introduce dev/prod profiles in webpack
