+++ babel.config.json
@@ -1,0 +1,8 @@
+{
+  "presets": [
+    "@babel/preset-env"
+  ],
+  "plugins": [
+    ["angularjs-annotate", { "es6": true, "dynamicImport": true, "explicitOnly": false }]
+  ]
+}
