+++ package.json
@@ -6,7 +6,9 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "serve": "webpack-dev-server --watch --progress",
-    "build:dev": "webpack"
+    "serve:prod": "webpack-dev-server --env=prod",
+    "build:dev": "webpack",
+    "build:prod": "webpack --env=prod"
   },
   "author": "",
   "license": "aGPLv3",
@@ -15,6 +17,12 @@
     "esn-frontend-mailto-handler": "github:openpaas-suite/esn-frontend-mailto-handler#main"
   },
   "devDependencies": {
+    "@babel/core": "^7.10.5",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+    "@babel/preset-env": "^7.10.4",
+    "babel-loader": "^8.1.0",
+    "babel-plugin-angularjs-annotate": "^0.10.0",
+    "clean-webpack-plugin": "^3.0.0",
     "css-loader": "^3.6.0",
     "exports-loader": "^1.0.0",
     "expose-loader": "^1.0.0",
@@ -28,9 +36,11 @@
     "pug-loader": "^2.4.0",
     "style-loader": "^1.2.1",
     "svg-inline-loader": "^0.8.2",
+    "terser-webpack-plugin": "^3.0.6",
     "url-loader": "^4.1.0",
     "webpack": "^4.43.0",
     "webpack-cli": "^3.3.11",
-    "webpack-dev-server": "^3.11.0"
+    "webpack-dev-server": "^3.11.0",
+    "webpack-merge": "^5.0.9"
   }
 }
